Update tagging.js
Since typeahead went to version 0.10.1 TagsManager can't get an ajax source. This was because Bloodhound and Typeahead are separate now. 

The above method properly encapsulates your TagsManager initialization and makes sure Bloodhound gets all the tags from the server and feeds them into each instance of typeahead.

For this to work i have modified widgets.html.twig. Nothing major, just so that the correct functions can be passed and used in the above.